refactor: components 하위 파일 구조 개선 	renamed:    src/components/Footer.css -> src/components/Footer/Footer.css 	renamed:    src/components/Footer.js -> src/components/Footer/Footer.js 	renamed:    src/components/Gnb.css -> src/components/Header/Gnb.css 	renamed:    src/components/Gnb.js -> src/components/Header/Gnb.js 	renamed:    src/components/Header.css -> src/components/Header/Header.css 	renamed:    src/components/Header.js -> src/components/Header/Header.js 	renamed:    src/components/SharedHeaderContent.css -> src/components/Header/SharedHeaderContent.css 	renamed:    src/components/SharedHeaderContent.js -> src/components/Header/SharedHeaderContent.js 	renamed:    src/components/Card.css -> src/components/Main/Card.css 	renamed:    src/components/Card.js -> src/components/Main/Card.js 	renamed:    src/components/CardList.css -> src/components/Main/CardList.css 	renamed:    src/components/CardList.js -> src/components/Main/CardList.js 	renamed:    src/components/Main.css -> src/components/Main/Main.css 	renamed:    src/components/Main.js -> src/components/Main/Main.js 	renamed:    src/components/SearchBar.css -> src/components/Main/SearchBar.css 	renamed:    src/components/SearchBar.js -> src/components/Main/SearchBar.js 	modified:   src/pages/Shared/Shared.js
