+++ src/components/Footer/Footer.js
@@ -1,8 +1,8 @@
 import "./Footer.css";
-import facebook from "../assets/images/akar-icons_facebook-fill.svg";
-import twitter from "../assets/images/akar-icons_twitter-fill.svg";
-import youtube from "../assets/images/akar-icons_youtube-fill.svg";
-import instagram from "../assets/images/ant-design_instagram-filled.svg";
+import facebook from "../../assets/images/akar-icons_facebook-fill.svg";
+import twitter from "../../assets/images/akar-icons_twitter-fill.svg";
+import youtube from "../../assets/images/akar-icons_youtube-fill.svg";
+import instagram from "../../assets/images/ant-design_instagram-filled.svg";
 
 function Footer() {
   return (
