+++ src/components/Main/CardList.js
@@ -1,14 +1,14 @@
 import "./CardList.css";
 import Card from "./Card";
-import card1 from "../assets/images/card-1.png";
-import card2 from "../assets/images/card-2.png";
-import card3 from "../assets/images/card-3.png";
-import card4 from "../assets/images/card-4.png";
-import card5 from "../assets/images/card-5.png";
-import card6 from "../assets/images/card-6.png";
-import card7 from "../assets/images/card-7.png";
-import card8 from "../assets/images/card-8.png";
-import card9 from "../assets/images/card-9.png";
+import card1 from "../../assets/images/card-1.png";
+import card2 from "../../assets/images/card-2.png";
+import card3 from "../../assets/images/card-3.png";
+import card4 from "../../assets/images/card-4.png";
+import card5 from "../../assets/images/card-5.png";
+import card7 from "../../assets/images/card-7.png";
+import card8 from "../../assets/images/card-8.png";
+import card6 from "../../assets/images/card-6.png";
+import card9 from "../../assets/images/card-9.png";
 
 function CardList() {
   return (
